fix build on non-linux platforms and add docs
