+++ examples/external_eventloop/README.md
@@ -1,5 +1,7 @@
 Example running an eframe application on an external eventloop.
 
+This allows you to run your eframe application alongside other windows and/or toolkits on the same event loop.
+
 ```sh
 cargo run -p external_eventloop
 ```
